refactor: 리팩토링 및 CornerstoneViewer의 초기화 및 도구 활성화 로직을 훅으로 분리
